Avoid problems with 64-bit time_t
The clock_gettime() calls are being handled by calling
syscall(__NR_clock_gettime, ...), which is not portable between systems
using 32-bit and 64-bit time_t.  This is being done to avoid having to
link agains librt.

So, use the standard function, and add a test to see if we can compile
a test without the library, including it otherwise.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>