Fix SIGSEGV caused by time_t casted to long on 32bit systems
32bit systems have time_t set to long long int while 64bit system have
time_t set to long int. This is problematic as in the busybox code this
is not handled correctly and we have some casted to long and some not
casted at all.

Some arch (found this problem on a mt7621) may be restrictive about casting
and crash with segmentation fault if time_t is cast to %ld instead of the
correct %lld.

This is the cause of https://github.com/RIPE-NCC/ripe-atlas-software-probe/issues/74

Use the correct type and cast every time_t to (unsigned long long) so that
eperd and condmv doesn't crash anymore and the measurement works correctly.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>