Fix EL8 sysusers creation
Oracle el8 (presumably all of RHEL) does not support the sysusers.conf
notion of specifying uid:gid for the ID field. This modifies the file
to use the more archaic way of creating a user and assigning them to
a group, with the difference that each user has a primary gid with the
same name as the user, in addition the the group specified for
ripe-atlas.