Fix bug with chown when systemd is disabled
During compile, if systemd is disabled, the build breaks unless
--disable-chown is given during configure.

Problem is caused when --disable-systemd is passed (or not detected)
which also disables the installation hook where the chown should be
executed. This then folds the chown into the generation of the
FIRMWARE_APPS_VERSION file, which is done at an early point in the
build where installation directories have not been created yet.

Reported-by: César de Tassis Filho <ctassisf@gmail.com>