refactor: moving from directories to separate Dockerfiles for GitLab CI
Note: docker build's argument is not only the directory of the Dockerfile
but also the directory of the "context". If you copy any files into the
container, that is where they are copied from. It *is* necessary.

As such, if you build with custom Dockerfile, but import from the repo.
Example from this commit:
- base command: `docker build`
- tag for image name: `-t ${DOCKER_IMAGE}`
- specify dockerfile: `-f .gitlab-docker/${OS}-Dockerfile`
- context: `.`