fix: OL10 Dockerfile for rpmlint
Oracle Linux 10 seems to have removed rpmlint from the default repositories.
This causes the `dnf` command to fail.

Instead, rpmlint has been moved to EPEL.

As to avoid contaminating the building process and final packages with
anything EPEL-related, splitting up the files keeps EPEL to the linting step.

Whether this fear is substantiated is unknown, but it does lower risk.

Preferably, we would make use of a single Dockerfile with multiple stages, but
it is unsure whether this is viable. This needs to be visited alongside kaniko.